% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaMatrix-covc.R, R/sigmaMatrix-sigmacap.R
\name{covc}
\alias{covc}
\alias{sigmacap}
\title{Consistent Estimate of the Covariance Matrix}
\usage{
covc(x)

sigmacap(x)
}
\arguments{
\item{x}{Numeric matrix, data frame, or vector.
Data.}
}
\description{
Consistent Estimate of the Covariance Matrix
}
\examples{
data(iris)
sigmacap(iris[, 1:4])
covc(iris[, 1:4])
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Covariance Functions}
\keyword{covariance}
\keyword{sigmaMatrix}
