% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaMatrix-package.R
\docType{package}
\name{sigmaMatrix}
\alias{sigmaMatrix}
\title{sigmaMatrix: Covariance Matrix}
\description{
A collection of functions related
to the covariance matrix \eqn{\boldsymbol{\Sigma}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{package}
\keyword{sigmaMatrix}
